+++ RELEASE_NOTES.md
@@ -1,1 +1,2 @@
-Command line options to run an update type only
+```--emacs``` to update emacs packages
+